101 refactored scripts and changed logic for more interactable for designers
