+++ Assets/Scripts/Player/PlayerDeathBehaviour.cs
@@ -1,11 +1,10 @@
 ﻿using System.Collections;
-using Player;
 using SceneMenager;
 using UnityEngine;
 using UnityEngine.Assertions;
 using UnityEngine.UI;
 
-namespace Sandbox.Vlad.Scripts
+namespace Player
 {
     public class PlayerDeathBehaviour : MonoBehaviour
     {
@@ -13,13 +12,15 @@ namespace Sandbox.Vlad.Scripts
         [SerializeField] private GameObject _deathCauseCanvas;
         private const float WaitBeforeReloadSceneSeconds = 5;
         private const float WaitBeforeFaideSeconds = 1.5f;
+        private AudioSource _deathRattle;
         private Animator _deathAnimation;
         private Image _faidingImage;
 
         private void Start()
         {
-            _deathAnimation = GetComponent<Animator>();
             _faidingImage = _deathCauseCanvas.GetComponent<Image>();
+            _deathAnimation = GetComponent<Animator>();
+            _deathRattle = GetComponent<AudioSource>();
         }
 
         private void StartPlayAnimation()
@@ -30,6 +31,19 @@ namespace Sandbox.Vlad.Scripts
 
         public IEnumerator ProcessOfDeath()
         {
+            PlayerBehaviour.PlayerDisable();
+            StartPlayAnimation();
+            yield return new WaitForSeconds(WaitBeforeFaideSeconds);
+            StartFaiding();
+            ShowTextOfDeathCause();
+            yield return new WaitForSeconds(WaitBeforeReloadSceneSeconds);
+            LoadScene.ReloadScene();
+        }
+        
+        public IEnumerator ProcessOfDeath(float secondsBeforeDeath)
+        {
+            yield return new WaitForSeconds(secondsBeforeDeath);
+            _deathRattle.Play();
             PlayerBehaviour.PlayerDisable();
             StartPlayAnimation();
             yield return new WaitForSeconds(WaitBeforeFaideSeconds);
@@ -51,7 +65,7 @@ namespace Sandbox.Vlad.Scripts
             fixedColor.a = 1;
             _faidingImage.color = fixedColor;
             _faidingImage.CrossFadeAlpha(0f, 0f, true);
-            _faidingImage.CrossFadeAlpha(1, 5, false);
+            _faidingImage.CrossFadeAlpha(1, 3, false);
         }
 
         public void ChangeTextOfDeathCause(string deathCause)
@@ -59,9 +73,10 @@ namespace Sandbox.Vlad.Scripts
             _deathCauseCanvas.GetComponentInChildren<Text>().text = deathCause;
         }
 
-        private void OnValidate()
-        {
-            Assert.IsNotNull(_deathCauseCanvas);
-        }
+//TODO why there is null but it isn't?        
+//        private void OnValidate()
+//        {
+//            Assert.IsNotNull(_deathCauseCanvas);
+//        }
     }
 }
